Updated minAge function to limit month/day options
The minAge option is an essential option for any of us who need to
restrict registration (as an example) to individuals thirteen years or
older for legal reasons.

Currently, the bday-picker plugin limits the year dropdown, but doesn't
adjust the available months or days. So anyone born in 1999 (for
example) who hasn't had their birthday yet can put in an age (12 years
and however many months) that is below the minimum age.

I've added code to the plugin that uses the existing functionality for
preventing future dates to also adjust for months and days below the
minimum age.

I added a fourth example to the index.html just to show the new
functionality in place. Try selecting the most recent year and check the
month and day dropdowns.